Refactor artisan authorization logic into a concern
